alterações

+++ app/Http/Controllers/Admin/PlanController.php
@@ -32,8 +32,6 @@ class PlanController extends Controller
 
     public function store(StoreUpdatePlan $request)
     {   
-        //dd($request->all());
-
         $this->repository->create($request->all());
 
         return redirect()->route('plans.index');
@@ -54,10 +52,19 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                        ->with(['details'])
+                        ->where('url', $url)
+                        ->first();
 
         if(!$plan)
             return redirect()->back();
+
+        if($plan->details->count() > 0){
+            return redirect()
+                    ->back()
+                    ->with('error', 'Existem detalhes vinculados a esse plano, portanto não é possivel deletar.');
+        }
         
         $plan->delete();
